chore: moved prism theme and codemirror styles into components
