+++ src/components/Preview/CodePreview/index.js
@@ -1,10 +1,10 @@
 import React from 'react'
-import CopyButton from '../CopyButton'
-import { useAppState } from '../../context'
+import CopyButton from '../../CopyButton'
+import { useAppState } from '../../../context'
 import styled from 'styled-components'
 
 import { PrismLight as SyntaxHighlighter } from 'react-syntax-highlighter'
-import prism from '../../lib/Prism/theme'
+import prism from './prism-theme'
 import html from 'react-syntax-highlighter/dist/esm/languages/prism/markup'
 
 const CodeEditor = styled.div`
